Updates firmware version to 1.3.0
Updates the RadPro WiFi Bridge firmware version to 1.3.0.

This change ensures the correct version is reported in the user-agent
header for outbound requests to Radmon, OpenSenseMap, and GMCMap,
as well as in the web UI and MQTT topic.

Adds diagnostic features, including LED status updates and USB debug log level control.